- nuevas extensiones en el gitignore para los archivos latex - en HFirstBest se agrego el simbolo [] para el delete de los arreglos - cambios en el rm del makefile - Implementacion de searchtabu tabusearch, lowerbound, buildS. Todavia falta terminar estas funciones y implementar diversificacion - Se agrego la estructura de dato Triple que posiblemente sera utilizada para la lista tabu - se comenzo a hacer el informe en latex - implementacion de funciones max y min en utils.cpp

+++ HBestBest.cpp
@@ -6,7 +6,7 @@ int HBestBest(int Tbin, Packing* items, int Hbin, int Wbin, int k){
   int n = items->binNum;
   long combs = comb((n-1),k);
   if (combs > 100000)
-    combs = combs*(0.4);
+    combs = combs* (long) (0.4);
   
   //Variables temporales de la heuristica
   Packing pack;
